Finished styling and logic for step 1

+++ Website/frontend/assets/js/controllers/enrollment/mobile/choosePhone.js
@@ -43,7 +43,7 @@
                                 return message.code.toLowerCase() == data.verifyEsnResponseCode.toLowerCase();
                             }).message);
                     }
-                } else if ($scope.getDevicesCount() > 0 && $scope.networkType != data.provider) {
+                } else if ($scope.getDevicesCount() > 0 && mobileEnrollmentService.selectedNetwork.value != data.provider) {
                     $scope.hasError = true;
                     if ($scope.networkType == 'GSM') {
                         $scope.cdmaIneligible = true;
@@ -82,8 +82,8 @@
         mobileEnrollmentService.editedDevice = item;
         //remove the device from the cart items array
         enrollmentCartService.removeDeviceFromCart(item);
-        enrollmentCartService.setActiveService(service);
-        enrollmentStepsService.setFlow('phone', false).setStep('phoneFlowDevices');
+        $scope.phoneOptions = item;
+        $scope.phoneVerified = true;
     };
 
     $scope.editDeviceNumber = function (service, item) {
@@ -108,25 +108,10 @@
         }
         //remove the device from the cart items array
         enrollmentCartService.removeDeviceFromCart(item);
-        //if there is 1 device left in the cart, go to data plan selection,
-        //if 0, go to device selection, otherwise stay at the same step
-        var devicesCount = enrollmentCartService.getDevicesCount();
-        var activeService = enrollmentCartService.getActiveService();
-        var serviceType = 'Mobile';
-        if (devicesCount == 0 && serviceType == 'Mobile') {
-            enrollmentStepsService.setFlow('mobile', false).setStep('phoneFlowDevices');
-            $scope.phoneVerified = false;
-            $scope.phoneOptions.imeiNumber = '';
-            scrollService.scrollTo('phoneFlowDevices', 0, 0, angular.noop);
-        } else if (devicesCount == 1) {
-            enrollmentCartService.setActiveService(service);
-            enrollmentStepsService.setFlow('mobile', false).setStep('phoneFlowPlans');
-        } else if (devicesCount > 1) {
-            //make a server call to update the cart with the correct devices
-            $timeout(function() { 
-                enrollmentService.setAccountInformation(); 
-            }, 50);
-        }
+        enrollmentStepsService.setFlow('mobile', false).setStep('phoneFlowDevices');
+        $scope.phoneVerified = false;
+        $scope.phoneOptions.imeiNumber = '';
+        scrollService.scrollTo('phoneFlowDevices', 0, 0, angular.noop);
     };
     
     $scope.isCartFull = function () {
@@ -410,6 +395,8 @@
                 imeiNumber: $scope.phoneOptions.imeiNumber,
                 simNumber: $scope.phoneOptions.simNumber,
                 iccidNumber: $scope.phoneOptions.iccidNumber,
+                activationCode: $scope.phoneOptions.activationCode,
+                phoneOs: $scope.phoneOptions.phoneOs,
                 transferInfo: ($scope.phoneOptions.transferInfo.type == 'new') ? null : $scope.phoneOptions.transferInfo,
                 lte: $scope.phoneOptions.supportsLte
             };
@@ -475,37 +462,31 @@
         })
     };
 
-    $scope.$watch(mobileEnrollmentService.getEditedDevice, function (device) {
-        if (device.id != null) {
-            if (device.id == '7') {
-                $scope.mobileEnrollment.phoneTypeTab = 'existing';
-            } else {
-                $scope.setSelectedPhone(device.device.id);
-            }
-        }
-    });
-
     /**
      * Complete the Choose Network Step
      * @return {[type]} [description]
      */
     $scope.completeStep = function () {  
-        if ($scope.phoneVerified) {
-            $scope.addDeviceToCart();
-        }
-        
-        var activeService = enrollmentCartService.getActiveService();
-        var dataPlan = _(activeService.offerInformationByType).where({ key: 'Mobile' }).first();
-        
-        // reset the dataplan if one is selected
-        if (typeof dataPlan != 'undefined' && dataPlan.value.offerSelections.length > 0){
-            enrollmentCartService.removeMobileOffers(activeService);
-        }
-        $scope.clearPhoneSelection();
-        $scope.phoneVerified = false;
-        enrollmentStepsService.setStep('phoneFlowPlans');
+        if ($scope.getDevicesCount() > 0 || $scope.phoneVerified) {
+            if ($scope.phoneVerified) {
+                $scope.addDeviceToCart();
+            }
+            
+            var activeService = enrollmentCartService.getActiveService();
+            var dataPlan = _(activeService.offerInformationByType).where({ key: 'Mobile' }).first();
+            
+            // reset the dataplan if one is selected
+            if (typeof dataPlan != 'undefined' && dataPlan.value.offerSelections.length > 0){
+                enrollmentCartService.removeMobileOffers(activeService);
+            }
+            $scope.clearPhoneSelection();
+            $scope.phoneVerified = false;
+            enrollmentStepsService.setStep('phoneFlowPlans');
 
-        $scope.mobileEnrollmentService.currentStepNumber = 2;
+            $scope.mobileEnrollmentService.currentStepNumber = 2;
+        } else if ($scope.phoneOptions.imeiNumber != '') {
+            $scope.verifyPhone()
+        }
     };
 
 }]);
